Files uploaded.
This project requires multiple instances i.e. the user info needs to be entered first, which will then create the user data database that stores the hashed PW and usernames. The Server will need to be running prior to running the client, the server will draw information from the user data database to ensure that the client has entered valid credentials and may enter.